Lektor (#1)
I got tired of dealing with how weird harp.js was. It's not maintained
anymore and I'm not really sold on it's plugin architecture. After
looking around, I decided to roll my own static site generator using
Flask-Frozen and a typical Flask app pattern. This quickly got out of
hand because I wanted to maintain my existing content structure and all
the "route" generation methods were brittle. This was extremely
frustrating until I remembered that the Flask creator had created his
own CMS that worked with flat files.

Enter Lektor. I decided to use this instead of Pelican because I think
the GUI is nifty (though, I'll probably never use it) and it seems like
it's really easy to follow some more CMS-y patterns (it has a flat file
"database" wrapper with querying). 90% of this conversion time was
converting the existing templates over to Jinja2 (which I was going to
have to do because a lot of the templates were Jade [or Pug or whatever
the hell they are calling it now] and I really hate Jade[1]). Also, it's
builtin Webpack plugin saved me a lot of time converting this and didn't
require me to use gulp, which is awesome!

Lektor is not perfect (yet) because it makes some very opinionated
assumptions about your content.

1. Lektor requires that all content files end with `.lr`. While this
   usually isn't an issue (I can write an ftplugin to switch it's
   rendering to markdown), this provides issues for things outside of
   Vim. Primarily, Editorial on iOS can read it as a text file but will
   not render it in it's sexy markdown editor. I'm gonna Tweet him and
   see what he thinks. If it's not possible, I'll, open an issue on the
   Lektor project.
2. While Lektor is Flask, and, thus can run the Werkzeug debugger on
   errors, Lektor disables this with it's own non-interactive error
   page. This was especially painful because their docs, while good,
   aren't complete enough for me to understand all of the data attached
   to the model or in the context or whatever. When I'm using Flask,
   I have a Jinja filter that will raise an exception in the template so
   that I can launch the debugger and play around. I have opened an
   [issue](lektor/lektor#311) about this and
   I hope it gets fixed, though I understand why it's don this way.
3. Wish it would work in Python 3, but that's okay.
4. I was hoping that Lektor would be more like Flask. For example, say
   I wanted to integrate Lektor into an existing Flask application. The
   usual pattern I'd follow is declare it as an extension and init it
   and use it as I felt appropriate. This could work/benefit in a few
   ways (at the same time, maybe):
   a. I decorate specific routes as "Lektor"ed a la Flask Frozen that
      just return a template. When Lektor is compiled it collects that
      route from it's "url map" and everything is good. I know there
      would be issues with this, but it would be cool. (I could also do
      this directly with Flask Frozen, but the DSL for Lektor is
      cooler.)
   b. A shared application config/babel/url_for/wsgi setup.
   c. Having the editor as a part of the app would be cool too (but
      lektor isn't explictly made to run live on the web, would still be
      cool).
   d. Conversely, I would have loved to write some routes directly to
      lektor in the Flask pattern.
   Obviously, Lektor !== Flask and that's fine, but I a boy can dream,
   right?

Overall, I'm really happy with Lektor and the overall state of my site
now. This transition took me about a day to do and I daresay that
transitioning other sites over from other static site generators will be
super easy.

[1] Let's talk Jade for a second. It's syntax, while pretty, is limiting
because you can't really do conditional attributes. Furthermore, for
whatever reason, it's Vim sytax is SLOW. For example, PHP syntax
highlighting in Vim is composed of the PHP syntax/plugins as well as
HTML, CSS, JS, and any other syntax/plugins that hooks into those (for
example, the JSX syntax will hook into the JS). Understandably, PHP can
be a bit slow to render in Vim with a complex file, but it's still
somewhat responsive. Jade, on the other had, when just rendering Jade
would be flat out unresponsive and a pain in the ass to deal with.

Okay, so why didn't I just switch to EJS, which harp.js supports as
well? Because harp.js only exposes it markdown rendering command to
Jade. So, say you want to generate slides with inline markdown? YOU'RE
STUCK WITH JADE YOU STUPID ASSHOLE. *harp.js cackles in your face**.

Now that I'm using Lektor, I'm able to intermix markdown and HTML
without having to use a renderer because Jinja is sane and Lektor goes
out of it's way to make your live easier.