Resume Updates (#2)
I made these changes a few months ago and deployed them live to my server. Unfortunately, somehow I didn't push up the changes to GitHub. As a result, when I went to update this over my vacation, the old styles/functionality returned and I had no way to revert it. Furthermore, somehow my Linux backup system didn't have a version of the site with this content.

In any case, the deed is done and I am going to sleep.