Slim Down Docker Image (#3)
I took a look at the Dockerfile and realized that a few of the layers could be merged together and that would make the final image size a bit smaller.

The original image size was 303 MB. That's a huge Docker image, but there really isn't a way around that, as I am slipping static assets into the image. This means that the image will continue to grow as I add more assets to the site. I might want to toss these assets on something like S3 at some point, but my site is extremely low traffic and Docker Hub's storage is free so whatever. What this does impact is local development of my Salt server setup (this will be open sourced once I get out of my own way and finish it), as I need to pull this image on each fresh VM build. It's easily the largest image that I have to pull, so any shrinking of the image is welcome.

In any case, after merging the package installation layers together, I ended up with an image size of ~280 MB, a savings of around ~20 MB. Not bad, but I can do better. I then merged the site compilation layers together, which make the image ~270 MB. At this point, I'm sad that the image isn't smaller so I start poking around in the container looking for things I can delete.

Alright, so `apt` lists are something people delete in Docker images? So, let's delete that. Lektor has to install Python packages in order to compile the website? I don't need those after I compile it, so let's nuke it. What else can I delete? What about the source of the website that I am copying into the container? Can I delete that and save some size? Not really, because the `COPY` command is it's own layer and deleting that directory in a subsequent `RUN` layer doesn't negate that layer at all. But doing that does make the end image smaller because Lektor has to install Node modules there and deleting them results in a huge image size reduction.

So, after all of this, the compressed image size is 241 MB, a reduction of 61 MB! Not bad for a half hour fix.

Which leads to the future: How can I make this image even smaller? The most obvious one to me is getting rid of the `COPY` command for the source website. What that means in practice is that I would need to compile the website outside of the Docker container and only copy the generated website into the container. This is easily doable with a simple makefile, but it would also defeat my entire use case of Docker, which is to contain all the dependencies and build processes inside of a container so I don't have to manage that on any of the machines consuming this website. I might circle back to this, because Lektor isn't that invasive and I suspect that it would yield great gains in image size.