Fix The Website (#7)
Background

This site has been malfunctioning for the past few months. I determined this to be because the Let's Encrypt container wasn't able to generate and verify certificates. I banged my head against this problem for about two months. Finally, a potential employer called me out on this and fixing this went to the top of my list.

So I sat down and walked through the error message I was seeing and realized that the `VIRTUAL_HOST` env var wasn't included `www.eligundry.com` and when Let's Encrypt went to verify it, it would error out because a virtual host entry wasn't being auto-magically created by the Nginx reverse proxy container. I added it and boom, it worked.

I usually detest magic in programming, but I make an exception for how the reverse proxy works for this site. All you need to do is set the `VIRTUAL_HOST` env var on your application container and add [jwilder/nginx-proxy](https://github.com/jwilder/nginx-proxy) to your setup. It will look at all the other containers for that `VIRTUAL_HOST` env var and template out one big config for all the upstreams. And if you want automatic Let's Encrypt renewal and certificate generation, add [JrCs/docker-letsencrypt-nginx-proxy-companion](https://github.com/JrCs/docker-letsencrypt-nginx-proxy-companion), which is a companion container that will hook into the reverse proxy to create all the certificates. This does mean that they both need to mount the local Docker socket as a volume, which is somewhat dangerous, but whatever, I don't need to maintain an nginx config.

Changes

* Fix the aforementioned Let's Encrypt Container
* Add my new headshot to the About page
* Update some copy on the About page.